requirements.txt
flask
flask-cors
customtkinter
Pillow
mysql-connector-python
